lspinfo: move to module
